+++ example/headers.ts
@@ -3,8 +3,9 @@ import KingWorld from '../src'
 import cookie from '../src/index'
 
 const app = new KingWorld()
-	.get('/', ({ responseHeaders }) => {
-		// responseHeaders.append('a', 'b')
+	.get('/', ({ status, responseHeaders }) => {
+		responseHeaders.append('a', 'b')
+		status(400)
 
 		return new Response('Ok')
 	})