+++ src/utils.ts
@@ -18,11 +18,13 @@ export const isPromise = <T>(
 export const clone = <T extends Object | any[] = Object | any[]>(
 	aObject: T
 ): T => {
-	const bObject: Record<string, any> = Array.isArray(aObject) ? [] : {}
+	const bObject: Record<string, any> = Array.isArray(aObject)
+		? []
+		: Object.create(null)
 
 	let value: Partial<T>
 	for (const key in aObject) {
-		value = aObject[key]
+		value = aObject[key] as any
 
 		bObject[key as any] =
 			typeof value === 'object' ? clone(value as T) : value
@@ -31,10 +33,33 @@ export const clone = <T extends Object | any[] = Object | any[]>(
 	return bObject as T
 }
 
-export const mapArrayObject = (value: [string, string][]) => {
-	const object: Record<string, string> = {}
+// export const splitOnce = (char: string, s: string) => {
+// 	const i = s.indexOf(char)
 
-	value.forEach(([key, value]) => (object[key] = value))
+// 	return i === -1 ? [s, ''] : [s.slice(0, i), s.slice(i + 1)]
+// }
 
-	return object
+export const getPath = (url: string): string => {
+	const queryIndex = url.indexOf('?')
+
+	return url.substring(
+		url.charCodeAt(0) === 47 ? 0 : url.indexOf('/', 11),
+		queryIndex === -1 ? url.length : queryIndex
+	)
 }
+
+export const getQuery = (url: string): string => {
+	const queryIndex = url.indexOf('?') + 1
+
+	return queryIndex ? url.substring(queryIndex) : ''
+}
+
+export const parseQuery = (search: string) =>
+	!search
+		? {}
+		: search.split('&').reduce((result, each) => {
+				const i = each.indexOf('=')
+				result[each.slice(0, i)] = each.slice(i + 1)
+
+				return result
+		  }, {} as Record<string, string>)