+++ src/utils.ts
@@ -6,9 +6,9 @@ export const concatArrayObject = <T>(a: T[], b: T | T[]): T[] => [
 ]
 
 export const mergeHook = (a: Hook, b?: Hook | RegisterHook): Hook<any> => ({
-	onRequest: concatArrayObject(a?.onRequest, b?.onRequest ?? []) ?? [],
-	transform: concatArrayObject(a?.transform, b?.transform ?? []) ?? [],
-	preHandler: concatArrayObject(a?.preHandler, b?.preHandler ?? []) ?? []
+	onRequest: concatArrayObject(a.onRequest, b?.onRequest ?? []) ?? [],
+	transform: concatArrayObject(a.transform, b?.transform ?? []) ?? [],
+	preHandler: concatArrayObject(a.preHandler, b?.preHandler ?? []) ?? []
 })
 
 export const isPromise = <T>(
@@ -16,21 +16,11 @@ export const isPromise = <T>(
 ): response is Promise<T> => response instanceof Promise
 
 export const clone = <T extends Object | any[] = Object | any[]>(
-	aObject: T
+	value: T
 ): T => {
-	const bObject: Record<string, any> = Array.isArray(aObject)
-		? []
-		: Object.create(null)
+	const [cloned] = [value] as [T]
 
-	let value: Partial<T>
-	for (const key in aObject) {
-		value = aObject[key] as any
-
-		bObject[key as any] =
-			typeof value === 'object' ? clone(value as T) : value
-	}
-
-	return bObject as T
+	return cloned
 }
 
 // export const splitOnce = (char: string, s: string) => {
@@ -43,23 +33,23 @@ export const getPath = (url: string): string => {
 	const queryIndex = url.indexOf('?')
 
 	return url.substring(
-		url.charCodeAt(0) === 47 ? 0 : url.indexOf('/', 11),
+		url.indexOf('/'),
 		queryIndex === -1 ? url.length : queryIndex
 	)
 }
 
-export const getQuery = (url: string): string => {
+export const mapQuery = (url: string): Record<string, string> => {
 	const queryIndex = url.indexOf('?') + 1
 
-	return queryIndex ? url.substring(queryIndex) : ''
+	return queryIndex
+		? url
+				.substring(queryIndex)
+				.split('&')
+				.reduce((result, each) => {
+					const i = each.indexOf('=')
+					result[each.slice(0, i)] = each.slice(i + 1)
+
+					return result
+				}, {} as Record<string, string>)
+		: {}
 }
-
-export const parseQuery = (search: string) =>
-	!search
-		? {}
-		: search.split('&').reduce((result, each) => {
-				const i = each.indexOf('=')
-				result[each.slice(0, i)] = each.slice(i + 1)
-
-				return result
-		  }, {} as Record<string, string>)